operator/examples/fogverse/requirements.txt
aiokafka==0.12.0
async-timeout==5.0.1
confluent-kafka==2.8.2
filelock==3.18.0
Fogverse @ git+https://github.com/FXDROS/fogverse.git@6ce1c19c22690451245ba76f5d1fa5104cc931c0
fsspec==2025.3.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
